Add query cancellation support to dataset queries
This commit adds query cancelling to the generic sql driver, Druid and
Presto. It cancels the future the query is running in which will
interrupt the query and allow any query cancellation logic to run. If
the database doesn't support query cancellation, or the necessary
information (i.e. query id) isn't available, the code will behave as
it did before.