Fix bigquery support for sorting by agg expressions [ci drivers]
BigQuery does not allown an expression in an order by clause and
instead wants name of the aliased expression from the select
clause. BQ requires this, but will break other databases (such as
Oracle). This commit will find any aggregations or aggregate
expressions and alias them in MBQL before they are convert to honeysql
structures. This allows BQ to sort by an aggregate expression.